Crate azalea_world

The Minecraft world representation used in Azalea.

Re-exports

pub use chunk_storage::Chunk;

pub use chunk_storage::ChunkStorage;

pub use chunk_storage::PartialChunkStorage;

pub use chunk_storage::Section;

Modules

chunk_storage

find_blocks

heightmap

iterators

Iterators for iterating over Minecraft blocks and chunks, based on prismarine-world’s iterators.

palette

Structs

BitStorage

A compact list of integers with the given number of bits per entry.

Instance

A world where the chunks are stored as weak pointers. This is used for shared worlds.

InstanceContainer

A container of Instances (aka worlds). Instances are stored as a Weak pointer here, so if no clients are using an instance it will be forgotten.

InstanceName

The name of the Instance (world) the entity is in. If two entities share the same world name, we assume they’re in the same instance.

MinecraftEntityId

An entity ID used by Minecraft. These are not guaranteed to be unique in shared worlds, that’s what Entity is for.

PartialEntityInfos

Keep track of certain metadatas that are only relevant for this partial world.

PartialInstance

PartialInstances are usually owned by clients, and hold strong references to chunks and entities in Instances.

Enums

MoveEntityError