azalea_world::chunk_storage
Function get_block_state_from_sections

pub fn get_block_state_from_sections(
    sections: &[Section],
    pos: &ChunkBlockPos,
    min_y: i32,
) -> Option<BlockState>

Get the block state at the given position from a list of sections. Returns None if the position is out of bounds.